fix: corrected classname for LoadingComponent
